+++ template/vite.config.ts
@@ -1,4 +1,3 @@
-import { cloudflare } from "@cloudflare/vite-plugin";
 import tailwindcss from "@tailwindcss/vite";
 import react from "@vitejs/plugin-react";
 import fs from "fs";
@@ -21,7 +20,6 @@ export default defineConfig({
     react({
       babel: { plugins: ["@react-dev-inspector/babel-plugin"] },
     }),
-    cloudflare(),
     tailwindcss(),
     tsconfigPaths(),
   ],