+++ packages/app/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @hypershape-ai/utils@0.0.2
+  - dev-server-utils@0.0.2
   - @hypershape-ai/api@0.0.1
 
 ## 0.0.1
@@ -13,4 +13,4 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @hypershape-ai/utils@0.0.1
+  - dev-server-utils@0.0.1